refactor: OnlineUserService와 GameRoomService는 Manager 네이밍으로 교체
