+++ backend/src/main/java/com/cgs/backend/websocket/service/game/GameUpdateService.java
@@ -1,7 +1,7 @@
 package com.cgs.backend.websocket.service.game;
 
 import com.cgs.backend.websocket.dto.game.GameUpdateMessage;
-import com.cgs.backend.websocket.service.room.GameRoomService;
+import com.cgs.backend.websocket.service.manager.GameRoomManager;
 import com.cgs.backend.websocket.util.WebSocketEndpoint;
 import com.cgs.backend.websocket.util.WebSocketUtils;
 import lombok.RequiredArgsConstructor;
@@ -12,10 +12,10 @@ import org.springframework.stereotype.Service;
 public class GameUpdateService {
 
     private final WebSocketUtils webSocketUtils;
-    private final GameRoomService gameRoomService;
+    private final GameRoomManager gameRoomManager;
 
     public void updateGame(GameUpdateMessage message) {
-        String opponentId = gameRoomService.getOpponent(message.getRoomId(), message.getUserId());
+        String opponentId = gameRoomManager.getOpponent(message.getRoomId(), message.getUserId());
         webSocketUtils.publishMessage(WebSocketEndpoint.gameUpdate(opponentId), message);
     }
 }
